# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Thingdex API",
    "description": "Household inventory API. See `/docs` for interactive examples and `/openapi.json` for machine-readable specs.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/relations/{relation_id}": {
      "patch": {
        "tags": [
          "relations"
        ],
        "summary": "Update Relation",
        "description": "Update relation metadata. Attach/detach endpoints own active state changes.",
        "operationId": "update_relation_v1_relations__relation_id__patch",
        "parameters": [
          {
            "name": "relation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Relation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRelationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRelationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "relations"
        ],
        "summary": "Delete Relation",
        "description": "Delete a relation (requires detach first for in-use relations).",
        "operationId": "delete_relation_v1_relations__relation_id__delete",
        "parameters": [
          {
            "name": "relation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Relation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/relations/{relation_id}/detach": {
      "post": {
        "tags": [
          "relations"
        ],
        "summary": "Detach Relation",
        "description": "Detach a child from a parent and place it in a location.",
        "operationId": "detach_relation_v1_relations__relation_id__detach_post",
        "parameters": [
          {
            "name": "relation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Relation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRelationDetach"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRelationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemRelationDetach": {
        "properties": {
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          }
        },
        "type": "object",
        "title": "ItemRelationDetach"
      },
      "ItemRelationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "parent_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Parent Item Id"
          },
          "child_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Child Item Id"
          },
          "relation_type": {
            "type": "string",
            "title": "Relation Type"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "parent_item_id",
          "child_item_id",
          "relation_type",
          "active",
          "created_at"
        ],
        "title": "ItemRelationOut"
      },
      "ItemRelationUpdate": {
        "properties": {
          "quantity": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ItemRelationUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}